+++ zenith-float-num/src/for_3rd/mod.rs
@@ -1,4 +1,6 @@
 #[cfg(feature = "serde")]
+mod codec;
+#[cfg(feature = "serde")]
 mod de;
 #[cfg(feature = "serde")]
 mod ser;